Update tests for standard_stages
Modify logic in tests for correct behaviour to match the statement
"make sure all stages in original mpm represented in output"; add a
similar test for the final case.

Correct a call to standard_stages with the wrong arguments - this was
causing an error to be thrown but not for the expected reason.

Be consistent with test matrices used for testing errors.